pyproject.toml
[tool.poetry]
name = "imapsprayer"
version = "0.1.0"
description = "A simple IMAP password sprayer"
authors = ["Mayk <7395296+yok4i@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8 || ^3.9"
PySocks = "^1.7.1"
urllib3 = "^1.26.9"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
